float

+++ app/models.py
@@ -35,7 +35,7 @@ class User_point_input_model(models.Model):
                default=now, editable=True,
                help_text='yyyy-mm-dd')
 
-    Hours_of_sleep = models.PositiveIntegerField(default=0) #accept decimals
+    Hours_of_sleep = models.FloatField(default=0) #accept decimals
     Water_100oz = models.BooleanField(default=False)
     clean_eating = models.BooleanField(default=False)
     workout_intensity = models.PositiveIntegerField(
@@ -169,7 +169,7 @@ class Point_model(models.Model):
             this date and before this date.
 
     """
-    sleep_points = models.PositiveIntegerField(default=0) #max_digits=5, decimal_places=2
+    sleep_points = models.FloatField(default=0) #max_digits=5, decimal_places=2
     date = models.DateField(default=now, editable=True)
     water_points = models.PositiveIntegerField(default=0)
     workout_points = models.PositiveIntegerField(default=0)
